Add ticket transaction

+++ app/src/main/java/com/example/unieventos/ui/screens/SignUpScreen.kt
@@ -175,7 +175,7 @@ fun SignUpForm(
             onClick = {
                 usersViewModel.createUser(
                     User(
-                        cedula,
+                        cedula.toInt(),
                         cedula,
                         nombre,
                         direccion,
